Transform panic in uint2 and uint4 sums under invalid accumulations into SQL-level error
This commit eliminates a panic raised when a sum aggregation of uint2 or uint4 is faced with a negative accumulation. Negative accumulations in reductions can arise due to, e.g., incomplete or invalid source data. The new strategy is to convert the invalid aggregation error into a query-level error instead of a panic by introduction of a map that performs appropriate casts and a projection. The change is carried out at the level of HIR-to-MIR lowering, since the casts may introduce errors and these will be later analyzed by transforms. Additionally, we give the optimizer the possibility to delay evaluation of the casts and projections to enable arrangement reuse. Finally, the commit adjusts the constant-folding behavior to output the same datum type as expected from rendering, thus making the casts sane in both evaluation alternatives.

+++ src/expr/src/relation/func.rs
@@ -460,7 +460,9 @@ where
     if datums.peek().is_none() {
         Datum::Null
     } else {
-        let x: u64 = datums.map(|d| u64::from(d.unwrap_uint16())).sum();
+        let x: u128 = datums.map(|d| u128::from(d.unwrap_uint16())).sum();
+        // Note that we return here a numeric, but the value will be downcast
+        // as part of reduce planning to a uint8.
         Datum::from(x)
     }
 }
@@ -473,7 +475,9 @@ where
     if datums.peek().is_none() {
         Datum::Null
     } else {
-        let x: u64 = datums.map(|d| u64::from(d.unwrap_uint32())).sum();
+        let x: u128 = datums.map(|d| u128::from(d.unwrap_uint32())).sum();
+        // Note that we return here a numeric, but the value will be downcast
+        // as part of reduce planning to a uint8.
         Datum::from(x)
     }
 }
